Use plain int, not int64, for Add method
For a nicer-to-use API, TDigest's Add method should accept an int, not
an int64. Under the covers, everything will still be an int64.

+++ tdigest_test.go
@@ -191,7 +191,7 @@ func TestQuantile(t *testing.T) {
 func TestAddValue(t *testing.T) {
 	type testcase struct {
 		value  float64
-		weight int64
+		weight int
 		want   []*centroid
 	}
